+++ README.md
@@ -32,7 +32,7 @@
 
 * 사다리
   - [x] 참여인원의 이름은 최대 5글자로 제한
-  - [ ] 참여인원의 이름은 콤마(,)로 구분
+  - [x] 참여인원의 이름은 콤마(,)로 구분
   - [ ] 사다리의 라인은 연속될 수 없다
 * 화면
   - [ ] 참여인원의 이름을 입력 받는다
